Switch SDFV away from singleton pattern
